fix: rename typing.py to avoid shadowing Python stdlib typing module
