+++ voltaire_bundler/gas/gas_manager_v7v8v9.py
@@ -9,7 +9,7 @@ from voltaire_bundler.bundle.exceptions import \
     (ExecutionException, ExecutionExceptionCode,
      ValidationException, ValidationExceptionCode)
 from voltaire_bundler.gas.gas_manager import GasManager, calculate_deposit_slot_index, deep_union
-from voltaire_bundler.typing import Address
+from voltaire_bundler.custom_types import Address
 from voltaire_bundler.user_operation.models import FailedOp, FailedOpWithRevert
 from voltaire_bundler.user_operation.user_operation_handler import \
     decode_failed_op_event, decode_failed_op_with_revert_event
